refactor(RVCV-154): usecaseImpl을 수정한다.
