+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/reviewproperty/application/usecase/request/UpdateContainerRequest.java
@@ -1,31 +1,22 @@
 package com.romanticpipe.reviewcanvas.domain.reviewproperty.application.usecase.request;
 
+import com.romanticpipe.reviewcanvas.reviewproperty.domain.Border;
+import com.romanticpipe.reviewcanvas.reviewproperty.domain.Padding;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 
 @Schema(name = "PostReviewContainerRequest", description = "Review Container 디자인 수정 요청")
 public record UpdateContainerRequest(@Schema(description = "가로", requiredMode = Schema.RequiredMode.REQUIRED)
 									 @NotBlank String width,
-									 @Schema(description = "내부여백 왼쪽", requiredMode = Schema.RequiredMode.REQUIRED)
-									 @NotBlank String paddingLeft,
-									 @Schema(description = "내부여백 오른쪽", requiredMode = Schema.RequiredMode.REQUIRED)
-									 @NotBlank String paddingRight,
-									 @Schema(description = "내부여백 위쪽", requiredMode = Schema.RequiredMode.REQUIRED)
-									 @NotBlank String paddingTop,
-									 @Schema(description = "내부여백 아래쪽", requiredMode = Schema.RequiredMode.REQUIRED)
-									 @NotBlank String paddingBottom,
+									 @Schema(description = "내부여백", requiredMode = Schema.RequiredMode.REQUIRED)
+									 @NotBlank Padding padding,
 									 @Schema(description = "배경", requiredMode = Schema.RequiredMode.REQUIRED)
 									 @NotBlank String background,
-									 @Schema(description = "테두리 왼쪽", requiredMode = Schema.RequiredMode.REQUIRED)
-									 @NotBlank String boarderLeft,
-									 @Schema(description = "테두리 오른쪽", requiredMode = Schema.RequiredMode.REQUIRED)
-									 @NotBlank String boarderRight,
-									 @Schema(description = "테두리 위쪽", requiredMode = Schema.RequiredMode.REQUIRED)
-									 @NotBlank String boarderTop,
-									 @Schema(description = "테두리 아래쪽", requiredMode = Schema.RequiredMode.REQUIRED)
-									 @NotBlank String boarderBottom,
+									 @Schema(description = "테두리", requiredMode = Schema.RequiredMode.REQUIRED)
+									 @NotBlank Border border,
 									 @Schema(description = "테두리 색상", requiredMode = Schema.RequiredMode.REQUIRED)
-									 @NotBlank String boarderColor,
+									 @NotBlank String borderColor,
 									 @Schema(description = "그림자", requiredMode = Schema.RequiredMode.REQUIRED)
 									 @NotBlank String shadow
 ) {
